+++ .eslintrc.json
@@ -23,6 +23,7 @@
                 "props": false
             }
         ],
+        "no-restricted-properties": 0,
         "no-underscore-dangle": 0,
         "object-shorthand": 0,
         "import/no-named-as-default-member": 0,
